Changed Feet Counters to long to guarantee they will never overflow.
Added logic to adjust feet counter offset if Mahlo meter is reset, or something else bad happens.
Added mapping maximum bow, skew, and elongation.
Many bug fixes

+++ MahloClient/Views/MyColors.cs
@@ -38,7 +38,7 @@ namespace MahloClient.Views
     public static void SetFeetColor(DataGridView grid, DataGridViewCellFormattingEventArgs args, IServiceSettings settings)
     {
       GreigeRoll roll = (GreigeRoll)grid.Rows[args.RowIndex].DataBoundItem;
-      int measuredLength = (int)args.Value;
+      long measuredLength = (long)args.Value;
 
       (args.CellStyle.BackColor, args.CellStyle.ForeColor) =
         roll.RollLength == 0 ? (grid.DefaultCellStyle.BackColor, grid.DefaultCellStyle.ForeColor) :
